Reorganize current handler implementation
