+++ axion/application/__init__.py
@@ -1,15 +1,5 @@
 from axion.application import base_path
-from axion.application import handler
-
-resolve_handler = handler.make
-InvalidHandlerError = handler.InvalidHandlerError
 
 get_base_path = base_path.make
 
-__all__ = [
-    # handler operationss
-    'resolve_handler',
-    'InvalidHandlerError',
-    # base path operations
-    'get_base_path',
-]
+__all__ = ('get_base_path', )
