+++ axion/oas/__init__.py
@@ -10,6 +10,7 @@ from axion.oas.model import (
     OASOperationId,
     OASParameter,
     OASParameterLocation,
+    OASParameterName,
     OASRequestBody,
     OASReservedHeaders,
     OASServer,
@@ -31,6 +32,7 @@ __all__ = [
     'OASContent',
     'OASParameter',
     'OASParameterLocation',
+    'OASParameterName',
     'OASMediaType',
     'OASReservedHeaders',
     'OASRequestBody',
