+++ axion/oas/functions.py
@@ -14,7 +14,7 @@ _PARAM_IN_TO_CLS_MAP = {
 _PARAM_CLS_TO_IN_MAP = {v: k for k, v in _PARAM_IN_TO_CLS_MAP.items()}
 
 
-@functools.lru_cache(maxsize=100)
+@functools.lru_cache()
 def operation_filter_parameters(
         operation: model.OASOperation,
         *types: model.OASParameterLocation,
@@ -34,6 +34,5 @@ def operation_filter_parameters(
     return list(filter(lambda p: isinstance(p, expected_types), operation.parameters))
 
 
-@functools.lru_cache(maxsize=100)
 def parameter_in(param: model.OASParameter) -> model.OASParameterLocation:
     return _PARAM_CLS_TO_IN_MAP[type(param)]
