+++ tests/test_oas_parameter.py
@@ -1,5 +1,6 @@
 import typing as t
 
+from multidict import istr
 import pytest
 
 from axion.oas import model
@@ -34,7 +35,7 @@ def test_path_param_resolve() -> None:
     assert path.example == 'Test'
 
     cmp_path = model.OASPathParameter(
-        name=model.OASParameterName('app_id'),
+        name=model.OASParameterName(istr('app_id')),
         schema=(
             model.OASStringType(
                 default=None,
@@ -57,7 +58,7 @@ def test_path_param_resolve() -> None:
 
     assert cmp_path == path
 
-    cmp_path.name = model.OASParameterName('other')
+    cmp_path.name = model.OASParameterName(istr('other'))
     assert not cmp_path == path
     assert cmp_path != path
